#210767 - StringIndexOutOfBoundsException: String index out of range: 803

+++ html.parser/src/org/netbeans/modules/html/parser/ElementsFactory.java
@@ -820,7 +820,7 @@ public class ElementsFactory {
 
         private CharSequence source;
         private int nameOffset;
-        private byte valueOffset2nameOffsetDiff;
+        private short valueOffset2nameOffsetDiff;
         private byte nameLen;
         private short valueLen;
 
@@ -838,7 +838,7 @@ public class ElementsFactory {
             this.source = source;
 
             this.nameOffset = nameOffset;
-            this.valueOffset2nameOffsetDiff = (byte) (valueOffset - nameOffset);
+            this.valueOffset2nameOffsetDiff = (short) (valueOffset - nameOffset);
 
             this.nameLen = nameLen;
             this.valueLen = valueLen;
